refactor(api): authMiddleware better-auth integration
