+++ api/src/auth.ts
@@ -3,6 +3,10 @@ import { drizzleAdapter } from "better-auth/adapters/drizzle";
 import { db } from "./db";
 import * as schema from "./db/schema";
 import { env } from "./env";
+import type { MiddlewareHandler } from "hono";
+import type { Env } from ".";
+import { findUserById } from "./db/queries/users";
+import { HTTPException } from "hono/http-exception";
 
 export const auth = betterAuth({
   database: drizzleAdapter(db, {
@@ -18,3 +22,30 @@ export const auth = betterAuth({
     },
   },
 });
+
+export const authMiddleware: MiddlewareHandler<Env> = async (c, next) => {
+  const session = await auth.api.getSession({ headers: c.req.raw.headers });
+
+  if (!session) {
+    throw new HTTPException(401);
+  }
+
+  c.set("user", session.user);
+  c.set("session", session.session);
+
+  c.set("getUserData", async () => {
+    const user = await findUserById(session.user.id);
+
+    // we shouldn't leave the request handlers handle this, we handle this ourselves here.
+    // below we are throwing an HTTPException - that's an acceptable way for middlewares to
+    // take control and stop the request cycle early. hopefully it's also acceptable for
+    // such methods here
+    if (!user) {
+      if (!user) throw new HTTPException(401);
+    }
+
+    return user;
+  });
+
+  await next();
+};
